mod: シミュレーション結果を配列に詰めて集計するみたいなことをするときにusizeのほうが便利なため、時間のデータ型をusizeに変更

+++ src/primitive/time/sim_time.rs
@@ -1,7 +1,7 @@
 use std::fmt::{Display, Formatter};
 use std::ops::{Add, AddAssign, Sub, SubAssign};
 
-pub type TimeTick = u64;
+pub type TimeTick = usize;
 
 #[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug)]
 pub struct SimTime(TimeTick);
